requirements.txt
langchain==0.1.16
langchain_openai==0.1.3
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
SQLAlchemy==2.0.19
psycopg2
langchain-community